dashboard/requirements.txt
Brotli==1.0.9
brotlipy==0.7.0
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
dash==2.9.1
dash-auth==2.0.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dill==0.3.6
diskcache==5.4.0
distlib==0.3.6
filelock==3.10.2
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Compress==1.13
fonttools==4.38.0
gunicorn==20.1.0
identify==2.5.21
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
multiprocess==0.70.14
munkres==1.1.4
nodeenv==1.7.0
numpy==1.24.2
orjson==3.8.9
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip==23.0.1
platformdirs==3.1.1
plotly==5.13.1
pre-commit==3.2.0
psutil==5.9.4
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
redis==3.5.3
requests==2.28.2
setuptools==67.4.0
six==1.16.0
tenacity==8.2.1
tornado==6.2
typing_extensions==4.5.0
ukkonen==1.0.1
unicodedata2==15.0.0
urllib3==1.26.14
virtualenv==20.21.0
Werkzeug==2.2.3
wheel==0.38.4
zipp==3.14.0
